requirements.txt
pytest
webdriver_manager
pytest-variables
pytest-selenium